sake_defconfig: Set USERFAULTFD to n
Not totally sure what kind of incompatibility it was triggering in
userspace since Zygote crashed during early boot and we never really
booted properly.

F DEBUG   : Abort message: 'Check failed: errno == ENOSYS || errno == EACCES userfaultfd: Invalid argument'

As of [1], this was enabled to maintain compatibility with FCM lvl 6,
however, even with ACK merged in, this config isn't referenced on any
of the msm-5.4 devices. Google does enable it on 5.10 but that's it.
This definitely needs a little more investigation but for now, we're
not building userfaultfd.o.

[1]: https://github.com/StatiXOS/android_kernel_asus_sm8350/commit/6c977edc0857c925a803a38405174ebd04a18f20#diff-964834bfc482b7aac5603b12ffc6e7a71a497f69c9307cfb1e560af48d5c9140R109

Change-Id: I62026ff91095f7da339c6102aa1d15de45c552e3